refactor: Remove useQueryCozy
Using directly useQuery from `cozy-client`
