+++ src/components/Home/Home.spec.jsx
@@ -5,6 +5,10 @@ import { render } from '@testing-library/react'
 import AppLike from 'test/components/AppLike'
 import Home from 'src/components/Home/Home'
 
+jest.mock('cozy-client/dist/hooks/useQuery', () => {
+  return jest.fn(() => ({ data: [] }))
+})
+
 const setup = (hasPapers = false) => {
   return render(
     <AppLike>
@@ -24,5 +28,11 @@ describe('Home components:', () => {
     const { getByText } = setup()
 
     expect(getByText('Add your personal documents'))
+  })
+
+  it('should display Empty text & Existing label when data exists', () => {
+    const { getByText } = setup(true)
+
+    expect(getByText('Existing'))
   })
 })
