+++ src/components/HomeWrapper/HomeWrapper.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
+import { useQuery, isQueryLoading } from 'cozy-client'
 import Spinner from 'cozy-ui/transpiled/react/Spinner'
 
 import { getAllQualificationLabel } from 'src/helpers/queries'
-import { useQueryCozy } from 'src/components/Hooks/useQueryCozy'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import StepperDialogWrapper from 'src/components/StepperDialogWrapper/StepperDialogWrapper'
 import Home from 'src/components/Home/Home'
@@ -11,12 +11,13 @@ import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
 
 const HomeWrapper = () => {
   const { isStepperDialogOpen } = useStepperDialog()
-  const { data: allPapersLabel, isQueryLoading } = useQueryCozy(
-    getAllQualificationLabel
+  const { data: allPapersLabel, ...rest } = useQuery(
+    getAllQualificationLabel.definition,
+    getAllQualificationLabel.options
   )
 
   return !isStepperDialogOpen ? (
-    isQueryLoading ? (
+    isQueryLoading(rest) ? (
       <Spinner
         size="xxlarge"
         className="u-flex u-flex-justify-center u-mt-2 u-h-5"
