+++ src/components/Placeholders/FeaturedPlaceholdersList.jsx
@@ -1,17 +1,20 @@
 import React, { useMemo } from 'react'
 
+import { useQuery } from 'cozy-client'
 import List from 'cozy-ui/transpiled/react/MuiCozyTheme/List'
 import ListSubheader from 'cozy-ui/transpiled/react/MuiCozyTheme/ListSubheader'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 
 import Placeholder from 'src/components/Placeholders/Placeholder'
-import { getAllQualificationLabel } from 'src/utils/queries'
-import { useQuery } from 'src/components/Hooks/useQuery'
-import { getFeaturedPlaceholders } from 'src/utils/getPlaceholders'
+import { getAllQualificationLabel } from 'src/helpers/queries'
+import { getFeaturedPlaceholders } from 'src/helpers/findPlaceholders'
 
 const FeaturedPlaceholdersList = () => {
   const { t } = useI18n()
-  const { data: allPapers } = useQuery(getAllQualificationLabel)
+  const { data: allPapers } = useQuery(
+    getAllQualificationLabel.definition,
+    getAllQualificationLabel.options
+  )
   const featuredPlaceholders = useMemo(
     () => getFeaturedPlaceholders(allPapers),
     [allPapers]
